requirements.txt
Django==1.8.9
MySQL-python==1.2.5
argparse==1.2.1
decorator==4.0.9
djangorestframework==3.3.3
ipdb==0.8.1
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.2.1
wsgiref==0.1.2
